requirements.txt
arch>=6.3.0
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.6.0
statsmodels>=0.14.0
typer>=0.12.0
python-dotenv>=1.0.0
